+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - New UTM Tags `utm_content` and `utm_term` plausible/analytics#515
 - If a session was started without a screen_size it is updated if an event with screen_size occurs
 - Added `LISTEN_IP` configuration parameter plausible/analytics#1189
+- The breakdown endpoint with the property query `property=event:goal` returns custom goal properties (within `props`)
 
 ### Fixed
 - UI fix where multi-line text in pills would not be underlined properly on small screens.